FTR-92 FTR-10 moved helper functions into separate file in Utils.kt
