+++ app/src/main/java/hr/foi/air/servicesync/ui/components/ReservationItem.kt
@@ -29,6 +29,8 @@ import com.example.compose.primaryLight
 import com.example.compose.surfaceContainerDark
 import com.example.compose.surfaceContainerLight
 import hr.foi.air.servicesync.R
+import hr.foi.air.servicesync.business.calculateDaysUntilReservation
+import hr.foi.air.servicesync.business.formatDate
 import java.util.concurrent.TimeUnit
 
 @Composable
@@ -103,15 +105,4 @@ fun ReservationItem(
             }
         }
     }
-}
-
-fun formatDate(timestamp: Long): String {
-    val sdf = java.text.SimpleDateFormat("dd.MM.yyyy HH:mm", java.util.Locale.getDefault())
-    return sdf.format(java.util.Date(timestamp))
-}
-
-fun calculateDaysUntilReservation(reservationDate: Long): Long {
-    val currentDate = System.currentTimeMillis()
-    val differenceInMillis = reservationDate - currentDate
-    return TimeUnit.MILLISECONDS.toDays(differenceInMillis)
 }
